Проект Tойота

+++ src/project/Customer.java
@@ -2,9 +2,9 @@ package project;
 
 public class Customer {
     private String name;
-    private int money;
+    private double money;
 
-    public Customer(String name, int money) {
+    public Customer(String name, double money) {
         this.name = name;
         this.money = money;
     }
@@ -17,11 +17,11 @@ public class Customer {
         this.name = name;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 }
